requirements.txt
transformers
torch
librosa
accelerate
soundfile
python-dotenv
